[buildcff2vf] Redesign buildcff2vf to be easier to integrate into fonttools
Add test case for Feature Variation support in the GSUB table. I am checking it in here because I also  edited the source ufo files for the dollar glyph to test the logic in buildcff2vf for making compatible charstrings. This is needed because of of the optimizations in makeotfexe that convert flat curves to lines, and which omit path-closing linteo's. Once we have transitioned to using the new buildcff2vf, I will check this test into the main branch as a  test case for Feature Variation support in the GSUB table.